Better Deadzone. Removed old code. Updated XCI
Added a deadzone variable that can be edited. Defaults to 0.12f.
Updated XCI dependency. This fixes the snapping of the rotation to each cardianl direction. The character can now rotate freely.
Changed getting the input from input.getaxis to input.getaxisraw, 
Removed the 2Dmode boolean and everything associated with it.
Removed unused lines
Changed the max jitter to be a public variable instead of a private
spread jitter now a private variable. 
This makes more sense now.