Fixed Selection bug, Button animations
Fixed a bug where a player spam alternates A and B in character selection, they would never be able to return to the main screen. Fixed.
Applying button animations to when they are selected
changed the hard blue tint on buttons to a very light red.
Turned a bool off in animator