Bullets raycast while moving, HitByBullet interface, Added nullcheck
Turned off the colliders on the bullets, wait does this still kill players
Added a null check for player status, looks for the dictionary only if it contains the key
Changed the bullet to do raycasts based on its current position to where it will be the next frame.  instaed of using SWEPT collision with a collider.
Added a HitByBullet interface, used for door and glass, these will implement how they will interact if they are hit by a bullet, takes a hit point and velocity as its parameters.