1 or 2 handed spots, bullets play all wall particles found,
Moved bomb character out of Resources/Character to WIP simply because it is a useless character right now.
Each character has 2 objects to mount weapons to, Move class automatically finds these, no need to set references.
Changed looking for polygoncolliders to just looking for the parent class which is Collider2D.
Weapons have a held sprite variable. The weapon will change to this whenver the weapon is held. will revert back to the default sprite whenever the gun isn't held.