Fixed being unable to swap with weapon on ground
Due to the new kinematic guns. Swapping guns doesn't work as the first collder it finds is the one the player is currently holding. Instead the overlap circle will get all colliders found. The player will remember what their last weapon was. This will let the player pickup a different weapon each time. However if there is only 1 weapon on the ground (which might be their previous weapon) the player will still be allowed to pick it up because it is the only choice.