Multiple Spawn points, Fixed bat. Audio stuff. Spatial blend audio, Randomized pitches
Added a different players spawn points.
Fixed the collider on the bat causing it to not be picked up properly.
A lot of audio stuff, 3D sound because it was wanted.
Boolean to randomize the pitch or not added to objects that use the randomized pitch.
Improved empty clip firing sound.