Turned off some animators. Can run full game loop
SpawnPoints must *ALWAYS* be in this Order:
2 Players, 
3 Players,
4 Players,
any other order and the map WILL break. **Subject to change in future to make it not follow this order. 
Fixed finding spawn points. There are currently no random spawn points.
After a round is won the players will spawn at 0,0,0.
Due to changes in spawn points, this is still in construction. Hopefully get some of it done tomorrow.
Turned off the animators and some UI stuf, just so a test build can be made. Doing the animations will be done later.
*Found a bug* You can drop a weapon through glass. Quickly fix this later.
Fixed GameManger not deleting itself after returning to title screen, resulting in the game breaking after a match is done.