Points now go red to blue. Fixed loading bugs. Removed some debug logs
Whenever the player wins a match, the point colours now go from red to blue.
Added the animation for the point screen coming in from the top.
Player UI panels is currently bugged. due to the new Shaders used for the world. A new alternative wiill have to be found.
Removed some Debug.Logs as it was no longer needed.
Removed the 4 FUCKING LINES THAT BROKE EVERYTHING .
Fixed the player spawning. The old system would only work if player were in order so
1 and 2
1 2 and 3
1 2 3 and 4
any other combination will break the spawning.
Removed a boolean check. 
Removed a lot of mInstance = Singleton.Instance to make the singletons safer. Removed the Gamemanger the ingamescene